resource_server/requirements.txt
Django==2.2.*
djangorestframework==3.11.0
djangorestframework-simplejwt
requests